requirements.txt

python-telegram-bot==20.7
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.3.0
aiohttp==3.9.1

